moving clients to generated-clients
