+++ tests/integration/suite/daprd/api/generated-clients/statestore/statestore.gen.go
